# docs/swagger.yaml
definitions:
  receiver.ContextAdd:
    properties:
      content:
        description: 配置内容
        type: string
      name:
        description: 配置名
        type: string
      remarks:
        description: 配置说明
        type: string
    required:
    - content
    - name
    type: object
  receiver.ContextUpdate:
    properties:
      content:
        description: 配置内容
        type: string
      name:
        description: 配置名
        type: string
    required:
    - content
    - name
    type: object
  receiver.StarAdd:
    properties:
      content:
        description: '"提交内容"'
        maxLength: 300
        type: string
      emailNeed:
        default: false
        description: '"是否需要寄送服务"'
        type: boolean
      name:
        description: '"署名"'
        maxLength: 10
        type: string
    required:
    - content
    type: object
  receiver.UserInfoUpdate:
    properties:
      name:
        description: '"昵称"'
        maxLength: 10
        type: string
    required:
    - name
    type: object
  response.Response:
    properties:
      code:
        description: 自定义错误码
        type: integer
      data:
        description: 数据
      msg:
        description: 信息
        type: string
    type: object
info:
  contact:
    email: eyesyeager@gmail.com
    name: 耶瞳
    url: http://space.eyesspace.top
  description: 耶瞳星空，来自星星的寄语
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 耶瞳星空
  version: "1.0"
paths:
  /context/contextAdd:
    post:
      consumes:
      - application/json
      description: 权限：Admin
      parameters:
      - description: 短token
        in: header
        name: sToken
        required: true
        type: string
      - description: 长token
        in: header
        name: lToken
        required: true
        type: string
      - description: 请求参数
        in: body
        name: receiver
        required: true
        schema:
          $ref: '#/definitions/receiver.ContextAdd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 添加配置
      tags:
      - context
  /context/contextUpdate:
    put:
      consumes:
      - application/json
      description: 权限：Admin；若表中name不存在则添加配置，已存在则修改配置
      parameters:
      - description: 短token
        in: header
        name: sToken
        required: true
        type: string
      - description: 长token
        in: header
        name: lToken
        required: true
        type: string
      - description: 请求参数
        in: body
        name: receiver
        required: true
        schema:
          $ref: '#/definitions/receiver.ContextUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 修改配置
      tags:
      - context
  /context/initSite:
    get:
      consumes:
      - application/json
      description: 权限：Visitor；该接口处理方式要求前端必须是服务端渲染，因为接口会明文返回密钥等信息
      parameters:
      - description: 短token
        in: header
        name: sToken
        type: string
      - description: 长token
        in: header
        name: lToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 网站信息初始化
      tags:
      - context
  /star/starAdd:
    post:
      consumes:
      - application/json
      description: 权限：User
      parameters:
      - description: 短token
        in: header
        name: sToken
        required: true
        type: string
      - description: 长token
        in: header
        name: lToken
        required: true
        type: string
      - description: 请求参数
        in: body
        name: receiver
        required: true
        schema:
          $ref: '#/definitions/receiver.StarAdd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 添加星星
      tags:
      - star
  /star/starGetById/{id}:
    get:
      consumes:
      - application/json
      description: 权限：Visitor
      parameters:
      - description: 短token
        in: header
        name: sToken
        type: string
      - description: 长token
        in: header
        name: lToken
        type: string
      - description: 寄语id
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 通过id获取星星
      tags:
      - star
  /star/starsGet:
    get:
      consumes:
      - application/json
      description: 权限：Visitor
      parameters:
      - description: 短token
        in: header
        name: sToken
        type: string
      - description: 长token
        in: header
        name: lToken
        type: string
      - description: 已读id密串
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 批量查询星星
      tags:
      - star
  /track/login:
    get:
      consumes:
      - application/json
      description: 权限：Visitor，无响应内容
      parameters:
      - description: 短token
        in: header
        name: sToken
        type: string
      - description: 长token
        in: header
        name: lToken
        type: string
      - description: 埋点数据加密包
        in: query
        name: package
        required: true
        type: string
      responses: {}
      security:
      - sToken,lToken: []
      summary: 登录埋点
      tags:
      - track
  /track/visit:
    get:
      consumes:
      - application/json
      description: 权限：Visitor，无响应内容
      parameters:
      - description: 短token
        in: header
        name: sToken
        type: string
      - description: 长token
        in: header
        name: lToken
        type: string
      - description: 埋点数据加密包
        in: query
        name: package
        required: true
        type: string
      responses: {}
      security:
      - sToken,lToken: []
      summary: 进站埋点
      tags:
      - track
  /user/userInfoGet:
    get:
      consumes:
      - application/json
      description: 权限：User
      parameters:
      - description: 短token
        in: header
        name: sToken
        required: true
        type: string
      - description: 长token
        in: header
        name: lToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 获取用户基本信息
      tags:
      - userThrift
  /user/userInfoUpdate:
    put:
      consumes:
      - application/json
      description: 权限：User
      parameters:
      - description: 短token
        in: header
        name: sToken
        required: true
        type: string
      - description: 长token
        in: header
        name: lToken
        required: true
        type: string
      - description: 请求参数
        in: body
        name: receiver
        required: true
        schema:
          $ref: '#/definitions/receiver.UserInfoUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 修改用户基本信息
      tags:
      - userThrift
  /user/userUpdateAvatar:
    post:
      consumes:
      - multipart/form-data
      description: 权限：User
      parameters:
      - description: 短token
        in: header
        name: sToken
        required: true
        type: string
      - description: 长token
        in: header
        name: lToken
        required: true
        type: string
      - description: 头像文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - sToken,lToken: []
      summary: 修改用户头像
      tags:
      - userThrift
swagger: "2.0"